Use refactored code and move to subdirectory
